refactor: move auth handlers
deps: add axum-jwt
