+++ Cargo.toml
@@ -51,6 +51,7 @@ axum-extra = { version = "0.10", features = [
 	"typed-header",
 	"file-stream",
 ] }
+axum-jwt = "=0.1.0" # ^0.1.1 uses Rust 1.88
 axum-test = "18"
 base64 = "0.21"
 bigdecimal = { version = "0.4.5", features = ["serde"] }
