+++ backend/lib/Cargo.toml
@@ -8,6 +8,7 @@ anyhow = { workspace = true }
 async-trait = { workspace = true }
 axum = { workspace = true }
 axum-extra = { workspace = true }
+axum-jwt = { workspace = true }
 base64 = { workspace = true }
 bigdecimal = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
