+++ backend/lib/src/api/routes.rs
@@ -21,11 +21,11 @@ pub fn routes(services: Services) -> Router {
 
     Router::new()
         // Auth routes
-        .route("/auth/nonce", post(handlers::nonce))
-        .route("/auth/verify", post(handlers::verify))
-        .route("/auth/refresh", post(handlers::refresh))
-        .route("/auth/logout", post(handlers::logout))
-        .route("/auth/profile", get(handlers::profile))
+        .route("/auth/nonce", post(handlers::auth::nonce))
+        .route("/auth/verify", post(handlers::auth::verify))
+        .route("/auth/refresh", post(handlers::auth::refresh))
+        .route("/auth/logout", post(handlers::auth::logout))
+        .route("/auth/profile", get(handlers::auth::profile))
         // MSP info routes
         .route("/info", get(handlers::info))
         .route("/stats", get(handlers::stats))
